requirements.txt
alembic==1.13.2
asgiref==3.8.1
Authlib==1.3.1
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.16.1
azure-mgmt-core==1.4.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.1.1
azure-mgmt-subscription==3.1.1
Babel==2.15.0
bcrypt==4.1.3
bidict==0.23.1
blinker==1.8.2
boto3==1.34.144
botocore==1.34.144
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
Django==5.0.7
djangorestframework==3.15.2
dnspython==2.6.1
drf-yasg==1.21.7
email_validator==2.2.0
eventlet==0.36.1
Flask==3.0.3
flask-babel==4.0.0
Flask-Compress==1.15
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.4.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
h11==0.14.0
httpagentparser==1.9.5
httplib2==0.22.0
idna==3.7
importlib_resources==6.4.0
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.4
jmespath==1.0.1
jsonformatter==0.3.2
keyring==24.3.1
ldap3==2.9.1
libgravatar==1.0.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.3.0
msal==1.29.0
msal-extensions==1.2.0
msrest==0.7.1
oauthlib==3.2.2
packaging==24.1
paramiko==3.4.0
passlib==1.7.4
pgadmin4==8.9
portalocker==2.10.1
proto-plus==1.24.0
protobuf==5.27.2
psutil==5.9.8
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
SecretStorage==3.3.3
setuptools==70.3.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
speaklater3==1.4
SQLAlchemy==2.0.31
sqlparse==0.5.0
sshtunnel==0.4.0
typer==0.12.3
typing_extensions==4.12.2
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.19
user-agents==2.2.0
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
zstandard==0.22.0
